:memo: Restructured the file structure after discussing it with team members
